Updated function entry points and manifest imports, removed obsolete files, and modified middleware secret. Adjusted renderer paths for consistency across the project.
