+++ .vercel/output/functions/_render.func/chunks/_astro_assets_7m0C5vIT.mjs
@@ -504,7 +504,7 @@ async function getConfiguredImageService() {
   if (!globalThis?.astroAsset?.imageService) {
     const { default: service } = await import(
       // @ts-expect-error
-      './sharp_CWDnBT1M.mjs'
+      './sharp_BDSZy8do.mjs'
     ).catch((e) => {
       const error = new AstroError(InvalidImageService);
       error.cause = e;
@@ -676,7 +676,7 @@ const $$Image = createComponent(async ($$result, $$props, $$slots) => {
   }
   const { class: className, ...attributes } = { ...additionalAttributes, ...image.attributes };
   return renderTemplate`${maybeRenderHead()}<img${addAttribute(image.src, "src")}${spreadAttributes(attributes)}${addAttribute(className, "class")}>`;
-}, "/Users/martinseibert/Documents/code/oranienschule-website/node_modules/.pnpm/astro@5.9.3_@types+node@24.0.3_rollup@4.43.0_typescript@5.8.3/node_modules/astro/components/Image.astro", void 0);
+}, "/Users/Carl-Axel/Documents/code/oranienschule/node_modules/.pnpm/astro@5.9.3_@types+node@24.0.3_rollup@4.43.0_typescript@5.8.3/node_modules/astro/components/Image.astro", void 0);
 
 const mimes = {
   "3g2": "video/3gpp2",
@@ -1194,7 +1194,7 @@ const $$Picture = createComponent(async ($$result, $$props, $$slots) => {
     const srcsetAttribute = props.densities || !props.densities && !props.widths && !useResponsive ? `${image.src}${image.srcSet.values.length > 0 ? ", " + image.srcSet.attribute : ""}` : image.srcSet.attribute;
     return renderTemplate`<source${addAttribute(srcsetAttribute, "srcset")}${addAttribute(lookup(image.options.format ?? image.src) ?? `image/${image.options.format}`, "type")}${spreadAttributes(sourceAdditionalAttributes)}>`;
   })}  <img${addAttribute(fallbackImage.src, "src")}${spreadAttributes(attributes)}${addAttribute(className, "class")}> </picture>`;
-}, "/Users/martinseibert/Documents/code/oranienschule-website/node_modules/.pnpm/astro@5.9.3_@types+node@24.0.3_rollup@4.43.0_typescript@5.8.3/node_modules/astro/components/Picture.astro", void 0);
+}, "/Users/Carl-Axel/Documents/code/oranienschule/node_modules/.pnpm/astro@5.9.3_@types+node@24.0.3_rollup@4.43.0_typescript@5.8.3/node_modules/astro/components/Picture.astro", void 0);
 
 const mod = /*#__PURE__*/Object.freeze(/*#__PURE__*/Object.defineProperty({
   __proto__: null
@@ -1217,7 +1217,7 @@ const $$Font = createComponent(($$result, $$props, $$slots) => {
     });
   }
   return renderTemplate`${preload && data.preloadData.map(({ url, type }) => renderTemplate`<link rel="preload"${addAttribute(url, "href")} as="font"${addAttribute(`font/${type}`, "type")} crossorigin>`)}<style>${unescapeHTML(data.css)}</style>`;
-}, "/Users/martinseibert/Documents/code/oranienschule-website/node_modules/.pnpm/astro@5.9.3_@types+node@24.0.3_rollup@4.43.0_typescript@5.8.3/node_modules/astro/components/Font.astro", void 0);
+}, "/Users/Carl-Axel/Documents/code/oranienschule/node_modules/.pnpm/astro@5.9.3_@types+node@24.0.3_rollup@4.43.0_typescript@5.8.3/node_modules/astro/components/Font.astro", void 0);
 
 const imageConfig = {"endpoint":{"route":"/_image"},"service":{"entrypoint":"astro/assets/services/sharp","config":{}},"domains":[],"remotePatterns":[],"experimentalDefaultStyles":true,"experimentalResponsiveImages":false};
 							const getImage = async (options) => await getImage$1(options, imageConfig);
