+++ .vercel/output/functions/_render.func/pages/spendeninformation.astro.mjs
@@ -1,0 +1,1 @@
+// Contents removed by Astro as it's used for prerendering only
