+++ .vercel/output/static/rss.xml
@@ -1,1 +1,1 @@
-<?xml version="1.0" encoding="UTF-8"?><rss version="2.0"><channel><title>Oranienschule Wiesbaden</title><description>Gymnasium der Landeshauptstadt Wiesbaden</description><link>https://example.com/</link><item><title>Using MDX</title><link>https://example.com/blog/using-mdx/</link><guid isPermaLink="true">https://example.com/blog/using-mdx/</guid><description>Lorem ipsum dolor sit amet</description><pubDate>Fri, 31 May 2024 22:00:00 GMT</pubDate></item><item><title>First post</title><link>https://example.com/blog/first-post/</link><guid isPermaLink="true">https://example.com/blog/first-post/</guid><description>Lorem ipsum dolor sit amet</description><pubDate>Thu, 07 Jul 2022 22:00:00 GMT</pubDate></item><item><title>Pull-out-Tage 2025: Vielfältiges Förderprogramm startet</title><link>https://example.com/blog/gymnasium-wiesbaden/</link><guid isPermaLink="true">https://example.com/blog/gymnasium-wiesbaden/</guid><description>Begabtenförderung an der Oranienschule mit spannenden Workshops vom 3D-Druck bis zur Fitnesslehre</description><pubDate>Sat, 08 Feb 2025 23:00:00 GMT</pubDate></item><item><title>Second post</title><link>https://example.com/blog/second-post/</link><guid isPermaLink="true">https://example.com/blog/second-post/</guid><description>Lorem ipsum dolor sit amet</description><pubDate>Thu, 14 Jul 2022 22:00:00 GMT</pubDate></item><item><title>Third post</title><link>https://example.com/blog/third-post/</link><guid isPermaLink="true">https://example.com/blog/third-post/</guid><description>Lorem ipsum dolor sit amet</description><pubDate>Thu, 21 Jul 2022 22:00:00 GMT</pubDate></item><item><title>Markdown Style Guide</title><link>https://example.com/blog/markdown-style-guide/</link><guid isPermaLink="true">https://example.com/blog/markdown-style-guide/</guid><description>Here is a sample of some basic Markdown syntax that can be used when writing Markdown content in Astro.</description><pubDate>Tue, 18 Jun 2024 22:00:00 GMT</pubDate></item></channel></rss>
+<?xml version="1.0" encoding="UTF-8"?><rss version="2.0"><channel><title>Oranienschule Wiesbaden</title><description>Gymnasium der Landeshauptstadt Wiesbaden</description><link>https://example.com/</link><item><title>Using MDX</title><link>https://example.com/blog/using-mdx/</link><guid isPermaLink="true">https://example.com/blog/using-mdx/</guid><description>Lorem ipsum dolor sit amet</description><pubDate>Fri, 31 May 2024 22:00:00 GMT</pubDate></item><item><title>Pull-out-Tage 2025: Vielfältiges Förderprogramm startet</title><link>https://example.com/blog/gymnasium-wiesbaden/</link><guid isPermaLink="true">https://example.com/blog/gymnasium-wiesbaden/</guid><description>Begabtenförderung an der Oranienschule mit spannenden Workshops vom 3D-Druck bis zur Fitnesslehre</description><pubDate>Sat, 08 Feb 2025 23:00:00 GMT</pubDate></item><item><title>First post</title><link>https://example.com/blog/first-post/</link><guid isPermaLink="true">https://example.com/blog/first-post/</guid><description>Lorem ipsum dolor sit amet</description><pubDate>Thu, 07 Jul 2022 22:00:00 GMT</pubDate></item><item><title>Third post</title><link>https://example.com/blog/third-post/</link><guid isPermaLink="true">https://example.com/blog/third-post/</guid><description>Lorem ipsum dolor sit amet</description><pubDate>Thu, 21 Jul 2022 22:00:00 GMT</pubDate></item><item><title>Second post</title><link>https://example.com/blog/second-post/</link><guid isPermaLink="true">https://example.com/blog/second-post/</guid><description>Lorem ipsum dolor sit amet</description><pubDate>Thu, 14 Jul 2022 22:00:00 GMT</pubDate></item><item><title>Markdown Style Guide</title><link>https://example.com/blog/markdown-style-guide/</link><guid isPermaLink="true">https://example.com/blog/markdown-style-guide/</guid><description>Here is a sample of some basic Markdown syntax that can be used when writing Markdown content in Astro.</description><pubDate>Tue, 18 Jun 2024 22:00:00 GMT</pubDate></item></channel></rss>
