+++ .vercel/output/static/sitemap-0.xml
@@ -1,1 +1,1 @@
-<?xml version="1.0" encoding="UTF-8"?><urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1"><url><loc>https://example.com/</loc></url><url><loc>https://example.com/about/</loc></url><url><loc>https://example.com/blog/</loc></url><url><loc>https://example.com/blog/first-post/</loc></url><url><loc>https://example.com/blog/gymnasium-wiesbaden/</loc></url><url><loc>https://example.com/blog/markdown-style-guide/</loc></url><url><loc>https://example.com/blog/second-post/</loc></url><url><loc>https://example.com/blog/third-post/</loc></url><url><loc>https://example.com/blog/using-mdx/</loc></url><url><loc>https://example.com/ContactForm/</loc></url></urlset>
+<?xml version="1.0" encoding="UTF-8"?><urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1"><url><loc>https://example.com/</loc></url><url><loc>https://example.com/about/</loc></url><url><loc>https://example.com/blog/</loc></url><url><loc>https://example.com/blog/first-post/</loc></url><url><loc>https://example.com/blog/gymnasium-wiesbaden/</loc></url><url><loc>https://example.com/blog/markdown-style-guide/</loc></url><url><loc>https://example.com/blog/second-post/</loc></url><url><loc>https://example.com/blog/third-post/</loc></url><url><loc>https://example.com/blog/using-mdx/</loc></url><url><loc>https://example.com/ContactForm/</loc></url><url><loc>https://example.com/spendeninformation/</loc></url></urlset>
